requirements.txt
alabaster==0.7.13
astroid==2.15.4
Babel==2.12.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
dill==0.3.6
docutils==0.18.1
flake8==6.0.0
idna==3.4
imagesize==1.4.1
isort==5.12.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.1
platformdirs==3.5.0
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.1
pylint==2.17.4
pyTelegramBotAPI==4.11.0
python-dotenv==1.0.0
requests==2.30.0
snowballstemmer==2.2.0
Sphinx==6.2.1
sphinx-rtd-theme==1.2.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
telebot==0.0.5
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.5.0
urllib3==2.0.2
wrapt==1.15.0
